+++ core/src/test/java/de/bitbrain/braingdx/tmx/MockTiledTileLayerBuilder.java
@@ -11,10 +11,15 @@ public class MockTiledTileLayerBuilder {
    private TiledMapTileLayer layer = Mockito.mock(TiledMapTileLayer.class);
 
    MockTiledTileLayerBuilder addCell(int x, int y) {
+      return addCell(x, y, false);
+   }
+
+   MockTiledTileLayerBuilder addCell(int x, int y, boolean collision) {
       Cell cell = Mockito.mock(Cell.class);
       TiledMapTile tile = Mockito.mock(TiledMapTile.class);
       MapProperties layerProperties = new MapProperties();
       MapProperties properties = new MapProperties();
+      properties.put(Constants.COLLISION, collision);
       Mockito.when(tile.getProperties()).thenReturn(properties);
       Mockito.when(cell.getTile()).thenReturn(tile);
       Mockito.when(layer.getCell(x, y)).thenReturn(cell);